requirements.txt
streamlit==1.41.1
openai==1.61.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
python-docx==1.1.2
Pillow==11.1.0
pdfplumber==0.11.4
